+++ packages/process/lib/Polysemy/Process/Interpreter/ProcessStdio.hs
@@ -21,10 +21,11 @@ import Polysemy.Process.Interpreter.Process (
 import Polysemy.Process.Interpreter.SystemProcess (PipesProcess, interpretSystemProcessNative)
 
 -- |Interpret 'Process' as a native 'Polysemy.Process.SystemProcess', producing unaccumulated chunks of 'ByteString'.
+-- Silently discards stderr.
 interpretProcessByteStringNative ::
   Members [Resource, Race, Async, Embed IO] r =>
-  -- |Whether to discard output chunks if the queue is full.
   ProcessOptions ->
+  -- |Basic config. The pipes will be changed to 'System.IO.Handle' by the interpreter.
   ProcessConfig () () () ->
   InterpreterFor (Scoped () (Process ByteString ByteString ByteString) !! ProcessError) r
 interpretProcessByteStringNative options conf =
@@ -33,6 +34,7 @@ interpretProcessByteStringNative options conf =
   raiseUnder
 
 -- |Interpret 'Process' as a native 'Polysemy.Process.SystemProcess', producing lines of 'ByteString'.
+-- Silently discards stderr.
 interpretProcessByteStringLinesNative ::
   Members [Resource, Race, Async, Embed IO] r =>
   ProcessOptions ->
@@ -45,24 +47,26 @@ interpretProcessByteStringLinesNative options conf =
   raiseUnder
 
 -- |Interpret 'Process' as a native 'Polysemy.Process.SystemProcess', producing unaccumulated chunks of 'Text'.
+-- Silently discards stderr.
 interpretProcessTextNative ::
   Members [Resource, Race, Async, Embed IO] r =>
   ProcessOptions ->
   -- |Basic config. The pipes will be changed to 'System.IO.Handle' by the interpreter.
   ProcessConfig () () () ->
-  InterpreterFor (Scoped () (Process ByteString Text Text) !! ProcessError) r
+  InterpreterFor (Scoped () (Process Text Text Text) !! ProcessError) r
 interpretProcessTextNative options conf =
   interpretSystemProcessNative conf .
   interpretProcessText @PipesProcess @SystemProcessError options .
   raiseUnder
 
 -- |Interpret 'Process' as a native 'Polysemy.Process.SystemProcess', producing lines of 'Text'.
+-- Silently discards stderr.
 interpretProcessTextLinesNative ::
   Members [Resource, Race, Async, Embed IO] r =>
   ProcessOptions ->
   -- |Basic config. The pipes will be changed to 'System.IO.Handle' by the interpreter.
   ProcessConfig () () () ->
-  InterpreterFor (Scoped () (Process ByteString Text Text) !! ProcessError) r
+  InterpreterFor (Scoped () (Process Text Text Text) !! ProcessError) r
 interpretProcessTextLinesNative options conf =
   interpretSystemProcessNative conf .
   interpretProcessTextLines @PipesProcess @SystemProcessError options .